Use TB protos where appropriate
TensorFlow is no longer exposing Python binding for proto definition.
TensorBoard has decided to copy all the required proto definitions over
and use it where possible.

We still rely on TensorFlow version of the proto for:
- TB summary v1 APIs. e.g., *.summary.pb()
- places where we use pywrap_tensorflow.EventsWriter

This commit mostly made mechanical changes for anything that
uses `tensor_util.make_tensor_proto` and anything that matches
`tf\.(Event|Summary)`.
