+++ tensorboard/backend/event_processing/db_import_multiplexer_test.py
@@ -21,6 +21,7 @@ import os
 import os.path
 import sqlite3
 
+from tensorboard.compat.proto.event_pb2 import Event
 from tensorboard.backend.event_processing import db_import_multiplexer
 from tensorboard.util import tensor_util
 from tensorboard.util import test_util
@@ -29,7 +30,7 @@ import tensorflow as tf
 
 def add_event(path):
   with test_util.FileWriterCache.get(path) as writer:
-    event = tf.Event()
+    event = Event()
     event.summary.value.add(tag='tag', tensor=tensor_util.make_tensor_proto(1))
     writer.add_event(event)
