+++ tensorboard/backend/event_processing/event_accumulator_test.py
@@ -25,7 +25,10 @@ from six.moves import xrange  # pylint: disable=redefined-builtin
 import tensorflow as tf
 
 from tensorboard.backend.event_processing import event_accumulator as ea
+from tensorboard.compat.proto.event_pb2 import Event, SessionLog
+from tensorboard.compat.proto.summary_pb2 import HistogramProto, Summary, SummaryMetadata
 from tensorboard.plugins.distribution import compressor
+from tensorboard.util import tensor_util
 from tensorboard.util import test_util
 
 
@@ -48,11 +51,11 @@ class _EventGenerator(object):
       yield self.items.pop(0)
 
   def AddScalar(self, tag, wall_time=0, step=0, value=0):
-    event = tf.Event(
+    event = Event(
         wall_time=wall_time,
         step=step,
-        summary=tf.Summary(
-            value=[tf.Summary.Value(tag=tag, simple_value=value)]))
+        summary=Summary(
+            value=[Summary.Value(tag=tag, simple_value=value)]))
     self.AddEvent(event)
 
   def AddHistogram(self,
@@ -66,7 +69,7 @@ class _EventGenerator(object):
                    hsum_squares=5,
                    hbucket_limit=None,
                    hbucket=None):
-    histo = tf.HistogramProto(
+    histo = HistogramProto(
         min=hmin,
         max=hmax,
         num=hnum,
@@ -74,10 +77,10 @@ class _EventGenerator(object):
         sum_squares=hsum_squares,
         bucket_limit=hbucket_limit,
         bucket=hbucket)
-    event = tf.Event(
+    event = Event(
         wall_time=wall_time,
         step=step,
-        summary=tf.Summary(value=[tf.Summary.Value(tag=tag, histo=histo)]))
+        summary=Summary(value=[Summary.Value(tag=tag, histo=histo)]))
     self.AddEvent(event)
 
   def AddImage(self,
@@ -87,12 +90,12 @@ class _EventGenerator(object):
                encoded_image_string=b'imgstr',
                width=150,
                height=100):
-    image = tf.Summary.Image(
+    image = Summary.Image(
         encoded_image_string=encoded_image_string, width=width, height=height)
-    event = tf.Event(
+    event = Event(
         wall_time=wall_time,
         step=step,
-        summary=tf.Summary(value=[tf.Summary.Value(tag=tag, image=image)]))
+        summary=Summary(value=[Summary.Value(tag=tag, image=image)]))
     self.AddEvent(event)
 
   def AddAudio(self,
@@ -103,15 +106,15 @@ class _EventGenerator(object):
                content_type='audio/wav',
                sample_rate=44100,
                length_frames=22050):
-    audio = tf.Summary.Audio(
+    audio = Summary.Audio(
         encoded_audio_string=encoded_audio_string,
         content_type=content_type,
         sample_rate=sample_rate,
         length_frames=length_frames)
-    event = tf.Event(
+    event = Event(
         wall_time=wall_time,
         step=step,
-        summary=tf.Summary(value=[tf.Summary.Value(tag=tag, audio=audio)]))
+        summary=Summary(value=[Summary.Value(tag=tag, audio=audio)]))
     self.AddEvent(event)
 
   def AddEvent(self, event):
@@ -455,7 +458,7 @@ class MockingEventAccumulatorTest(EventAccumulatorTest):
     gen = _EventGenerator(self)
     acc = ea.EventAccumulator(gen)
     gen.AddScalar('s1', wall_time=1, step=10, value=20)
-    gen.AddEvent(tf.Event(wall_time=2, step=20, file_version='nots2'))
+    gen.AddEvent(Event(wall_time=2, step=20, file_version='nots2'))
     gen.AddScalar('s3', wall_time=3, step=100, value=1)
     gen.AddHistogram('hst1')
     gen.AddImage('im1')
@@ -486,7 +489,7 @@ class MockingEventAccumulatorTest(EventAccumulatorTest):
     gen = _EventGenerator(self)
     acc = ea.EventAccumulator(gen)
 
-    gen.AddEvent(tf.Event(wall_time=0, step=0, file_version='brain.Event:1'))
+    gen.AddEvent(Event(wall_time=0, step=0, file_version='brain.Event:1'))
     gen.AddScalar('s1', wall_time=1, step=100, value=20)
     gen.AddScalar('s1', wall_time=1, step=200, value=20)
     gen.AddScalar('s1', wall_time=1, step=300, value=20)
@@ -507,7 +510,7 @@ class MockingEventAccumulatorTest(EventAccumulatorTest):
     gen = _EventGenerator(self)
     acc = ea.EventAccumulator(gen, purge_orphaned_data=False)
 
-    gen.AddEvent(tf.Event(wall_time=0, step=0, file_version='brain.Event:1'))
+    gen.AddEvent(Event(wall_time=0, step=0, file_version='brain.Event:1'))
     gen.AddScalar('s1', wall_time=1, step=100, value=20)
     gen.AddScalar('s1', wall_time=1, step=200, value=20)
     gen.AddScalar('s1', wall_time=1, step=300, value=20)
@@ -539,7 +542,7 @@ class MockingEventAccumulatorTest(EventAccumulatorTest):
     gen = _EventGenerator(self)
     acc = ea.EventAccumulator(gen)
 
-    gen.AddEvent(tf.Event(wall_time=0, step=0, file_version='brain.Event:1'))
+    gen.AddEvent(Event(wall_time=0, step=0, file_version='brain.Event:1'))
     gen.AddScalar('s1', wall_time=1, step=100, value=20)
     gen.AddScalar('s1', wall_time=1, step=200, value=20)
     gen.AddScalar('s1', wall_time=1, step=300, value=20)
@@ -564,9 +567,9 @@ class MockingEventAccumulatorTest(EventAccumulatorTest):
     gen = _EventGenerator(self)
     acc = ea.EventAccumulator(gen)
     gen.AddScalar('s1', wall_time=1, step=100, value=20)
-    ev1 = tf.Event(wall_time=2, step=0, file_version='brain.Event:1')
+    ev1 = Event(wall_time=2, step=0, file_version='brain.Event:1')
     graph_bytes = tf.GraphDef().SerializeToString()
-    ev2 = tf.Event(wall_time=3, step=0, graph_def=graph_bytes)
+    ev2 = Event(wall_time=3, step=0, graph_def=graph_bytes)
     gen.AddEvent(ev1)
     gen.AddEvent(ev2)
     acc.Reload()
@@ -581,7 +584,7 @@ class MockingEventAccumulatorTest(EventAccumulatorTest):
     """
     gen = _EventGenerator(self)
     acc = ea.EventAccumulator(gen)
-    gen.AddEvent(tf.Event(wall_time=0, step=1, file_version='brain.Event:2'))
+    gen.AddEvent(Event(wall_time=0, step=1, file_version='brain.Event:2'))
 
     gen.AddScalar('s1', wall_time=1, step=100, value=20)
     gen.AddScalar('s1', wall_time=1, step=200, value=20)
@@ -591,8 +594,8 @@ class MockingEventAccumulatorTest(EventAccumulatorTest):
     gen.AddScalar('s2', wall_time=1, step=202, value=20)
     gen.AddScalar('s2', wall_time=1, step=203, value=20)
 
-    slog = tf.SessionLog(status=tf.SessionLog.START)
-    gen.AddEvent(tf.Event(wall_time=2, step=201, session_log=slog))
+    slog = SessionLog(status=tf.SessionLog.START)
+    gen.AddEvent(Event(wall_time=2, step=201, session_log=slog))
     acc.Reload()
     self.assertEqual([x.step for x in acc.Scalars('s1')], [100, 200])
     self.assertEqual([x.step for x in acc.Scalars('s2')], [])
@@ -601,7 +604,7 @@ class MockingEventAccumulatorTest(EventAccumulatorTest):
     """Test that FirstEventTimestamp() returns wall_time of the first event."""
     gen = _EventGenerator(self)
     acc = ea.EventAccumulator(gen)
-    gen.AddEvent(tf.Event(wall_time=10, step=20, file_version='brain.Event:2'))
+    gen.AddEvent(Event(wall_time=10, step=20, file_version='brain.Event:2'))
     gen.AddScalar('s1', wall_time=30, step=40, value=20)
     self.assertEqual(acc.FirstEventTimestamp(), 10)
 
@@ -609,7 +612,7 @@ class MockingEventAccumulatorTest(EventAccumulatorTest):
     """Test that Reload() means FirstEventTimestamp() won't load events."""
     gen = _EventGenerator(self)
     acc = ea.EventAccumulator(gen)
-    gen.AddEvent(tf.Event(wall_time=1, step=2, file_version='brain.Event:2'))
+    gen.AddEvent(Event(wall_time=1, step=2, file_version='brain.Event:2'))
 
     acc.Reload()
 
@@ -623,7 +626,7 @@ class MockingEventAccumulatorTest(EventAccumulatorTest):
     """Test that FirstEventTimestamp() doesn't discard the loaded event."""
     gen = _EventGenerator(self)
     acc = ea.EventAccumulator(gen)
-    gen.AddEvent(tf.Event(wall_time=1, step=2, file_version='brain.Event:2'))
+    gen.AddEvent(Event(wall_time=1, step=2, file_version='brain.Event:2'))
 
     self.assertEqual(acc.FirstEventTimestamp(), 1)
     acc.Reload()
@@ -739,8 +742,8 @@ class RealisticEventAccumulatorTest(EventAccumulatorTest):
     """Test accumulator by writing values and then reading them."""
 
     def FakeScalarSummary(tag, value):
-      value = tf.Summary.Value(tag=tag, simple_value=value)
-      summary = tf.Summary(value=[value])
+      value = Summary.Value(tag=tag, simple_value=value)
+      summary = Summary(value=[value])
       return summary
 
     directory = os.path.join(self.get_temp_dir(), 'values_dir')
@@ -852,9 +855,9 @@ class RealisticEventAccumulatorTest(EventAccumulatorTest):
         same file
     """
 
-    summary = tf.Summary()
+    summary = Summary()
     summary.value.add(
-        tensor=tf.compat.v1.make_tensor_proto(['po', 'ta', 'to'], dtype=tf.string),
+        tensor=tensor_util.make_tensor_proto(['po', 'ta', 'to'], dtype=tf.string),
         tag='you_are_it',
         metadata=summary_metadata)
     writer = test_util.FileWriter(logdir, filename_suffix=nonce)
@@ -863,7 +866,7 @@ class RealisticEventAccumulatorTest(EventAccumulatorTest):
 
   def testSummaryMetadata(self):
     logdir = self.get_temp_dir()
-    summary_metadata = tf.SummaryMetadata(
+    summary_metadata = SummaryMetadata(
         display_name='current tagee', summary_description='no')
     summary_metadata.plugin_data.plugin_name = 'outlet'
     self._writeMetadata(logdir, summary_metadata)
@@ -874,18 +877,18 @@ class RealisticEventAccumulatorTest(EventAccumulatorTest):
 
   def testSummaryMetadata_FirstMetadataWins(self):
     logdir = self.get_temp_dir()
-    summary_metadata_1 = tf.SummaryMetadata(
+    summary_metadata_1 = SummaryMetadata(
         display_name='current tagee',
         summary_description='no',
-        plugin_data=tf.SummaryMetadata.PluginData(plugin_name='outlet',
+        plugin_data=SummaryMetadata.PluginData(plugin_name='outlet',
                                                   content=b'120v'))
     self._writeMetadata(logdir, summary_metadata_1, nonce='1')
     acc = ea.EventAccumulator(logdir)
     acc.Reload()
-    summary_metadata_2 = tf.SummaryMetadata(
+    summary_metadata_2 = SummaryMetadata(
         display_name='tagee of the future',
         summary_description='definitely not',
-        plugin_data=tf.SummaryMetadata.PluginData(plugin_name='plug',
+        plugin_data=SummaryMetadata.PluginData(plugin_name='plug',
                                                   content=b'110v'))
     self._writeMetadata(logdir, summary_metadata_2, nonce='2')
     acc.Reload()
@@ -898,18 +901,18 @@ class RealisticEventAccumulatorTest(EventAccumulatorTest):
     # set of plugins in the `plugin_data` of second is different from
     # that of the first, then the second set should be ignored.
     logdir = self.get_temp_dir()
-    summary_metadata_1 = tf.SummaryMetadata(
+    summary_metadata_1 = SummaryMetadata(
         display_name='current tagee',
         summary_description='no',
-        plugin_data=tf.SummaryMetadata.PluginData(plugin_name='outlet',
+        plugin_data=SummaryMetadata.PluginData(plugin_name='outlet',
                                                   content=b'120v'))
     self._writeMetadata(logdir, summary_metadata_1, nonce='1')
     acc = ea.EventAccumulator(logdir)
     acc.Reload()
-    summary_metadata_2 = tf.SummaryMetadata(
+    summary_metadata_2 = SummaryMetadata(
         display_name='tagee of the future',
         summary_description='definitely not',
-        plugin_data=tf.SummaryMetadata.PluginData(plugin_name='plug',
+        plugin_data=SummaryMetadata.PluginData(plugin_name='plug',
                                                   content=b'110v'))
     self._writeMetadata(logdir, summary_metadata_2, nonce='2')
     acc.Reload()
