+++ tensorboard/backend/event_processing/event_file_inspector_test.py
@@ -23,6 +23,8 @@ import shutil
 import tensorflow as tf
 
 from tensorboard.backend.event_processing import event_file_inspector as efi
+from tensorboard.compat.proto.event_pb2 import SessionLog
+from tensorboard.compat.proto.summary_pb2 import HistogramProto, Summary
 from tensorboard.util import test_util
 
 
@@ -48,13 +50,13 @@ class EventFileInspectorTest(tf.test.TestCase):
 
       with test_util.FileWriterCache.get(subdir) as sw:
         for datum in data:
-          summary = tf.Summary()
+          summary = Summary()
           if 'simple_value' in datum:
             summary.value.add(tag=datum['tag'],
                               simple_value=datum['simple_value'])
             sw.add_summary(summary, global_step=datum['step'])
           elif 'histo' in datum:
-            summary.value.add(tag=datum['tag'], histo=tf.HistogramProto())
+            summary.value.add(tag=datum['tag'], histo=HistogramProto())
             sw.add_summary(summary, global_step=datum['step'])
           elif 'session_log' in datum:
             sw.add_session_log(datum['session_log'], global_step=datum['step'])
@@ -123,31 +125,31 @@ class EventFileInspectorTest(tf.test.TestCase):
   def testSessionLogSummaries(self):
     data = [
         {
-            'session_log': tf.SessionLog(status=tf.SessionLog.START),
+            'session_log': SessionLog(status=SessionLog.START),
             'step': 0
         },
         {
-            'session_log': tf.SessionLog(status=tf.SessionLog.CHECKPOINT),
+            'session_log': SessionLog(status=SessionLog.CHECKPOINT),
             'step': 1
         },
         {
-            'session_log': tf.SessionLog(status=tf.SessionLog.CHECKPOINT),
+            'session_log': SessionLog(status=SessionLog.CHECKPOINT),
             'step': 2
         },
         {
-            'session_log': tf.SessionLog(status=tf.SessionLog.CHECKPOINT),
+            'session_log': SessionLog(status=SessionLog.CHECKPOINT),
             'step': 3
         },
         {
-            'session_log': tf.SessionLog(status=tf.SessionLog.STOP),
+            'session_log': SessionLog(status=SessionLog.STOP),
             'step': 4
         },
         {
-            'session_log': tf.SessionLog(status=tf.SessionLog.START),
+            'session_log': SessionLog(status=SessionLog.START),
             'step': 5
         },
         {
-            'session_log': tf.SessionLog(status=tf.SessionLog.STOP),
+            'session_log': SessionLog(status=SessionLog.STOP),
             'step': 6
         },
     ]
