+++ tensorboard/backend/event_processing/event_multiplexer.py
@@ -440,7 +440,7 @@ class EventMultiplexer(object):
         the given run.
 
     Returns:
-      A `tf.SummaryMetadata` protobuf.
+      A `SummaryMetadata` protobuf.
     """
     accumulator = self.GetAccumulator(run)
     return accumulator.SummaryMetadata(tag)
