+++ tensorboard/backend/event_processing/plugin_event_accumulator_test.py
@@ -25,6 +25,8 @@ from six.moves import xrange  # pylint: disable=redefined-builtin
 import tensorflow as tf
 
 from tensorboard.backend.event_processing import plugin_event_accumulator as ea
+from tensorboard.compat.proto.event_pb2 import Event, SessionLog
+from tensorboard.compat.proto.summary_pb2 import Summary, SummaryMetadata
 from tensorboard.plugins.audio import summary as audio_summary
 from tensorboard.plugins.image import summary as image_summary
 from tensorboard.plugins.scalar import summary as scalar_summary
@@ -58,11 +60,11 @@ class _EventGenerator(object):
     important.
     """
     tensor = tensor_util.make_tensor_proto(float(value))
-    event = tf.Event(
+    event = Event(
         wall_time=wall_time,
         step=step,
-        summary=tf.Summary(
-            value=[tf.Summary.Value(tag=tag, tensor=tensor)]))
+        summary=Summary(
+            value=[Summary.Value(tag=tag, tensor=tensor)]))
     self.AddEvent(event)
 
   def AddEvent(self, event):
@@ -164,7 +166,7 @@ class MockingEventAccumulatorTest(EventAccumulatorTest):
     gen = _EventGenerator(self)
     acc = ea.EventAccumulator(gen)
     gen.AddScalarTensor('s1', wall_time=1, step=10, value=20)
-    gen.AddEvent(tf.Event(wall_time=2, step=20, file_version='nots2'))
+    gen.AddEvent(Event(wall_time=2, step=20, file_version='nots2'))
     gen.AddScalarTensor('s3', wall_time=3, step=100, value=1)
 
     acc.Reload()
@@ -188,7 +190,7 @@ class MockingEventAccumulatorTest(EventAccumulatorTest):
     gen = _EventGenerator(self)
     acc = ea.EventAccumulator(gen)
 
-    gen.AddEvent(tf.Event(wall_time=0, step=0, file_version='brain.Event:1'))
+    gen.AddEvent(Event(wall_time=0, step=0, file_version='brain.Event:1'))
     gen.AddScalarTensor('s1', wall_time=1, step=100, value=20)
     gen.AddScalarTensor('s1', wall_time=1, step=200, value=20)
     gen.AddScalarTensor('s1', wall_time=1, step=300, value=20)
@@ -209,7 +211,7 @@ class MockingEventAccumulatorTest(EventAccumulatorTest):
     gen = _EventGenerator(self)
     acc = ea.EventAccumulator(gen, purge_orphaned_data=False)
 
-    gen.AddEvent(tf.Event(wall_time=0, step=0, file_version='brain.Event:1'))
+    gen.AddEvent(Event(wall_time=0, step=0, file_version='brain.Event:1'))
     gen.AddScalarTensor('s1', wall_time=1, step=100, value=20)
     gen.AddScalarTensor('s1', wall_time=1, step=200, value=20)
     gen.AddScalarTensor('s1', wall_time=1, step=300, value=20)
@@ -241,7 +243,7 @@ class MockingEventAccumulatorTest(EventAccumulatorTest):
     gen = _EventGenerator(self)
     acc = ea.EventAccumulator(gen)
 
-    gen.AddEvent(tf.Event(wall_time=0, step=0, file_version='brain.Event:1'))
+    gen.AddEvent(Event(wall_time=0, step=0, file_version='brain.Event:1'))
     gen.AddScalarTensor('s1', wall_time=1, step=100, value=20)
     gen.AddScalarTensor('s2', wall_time=1, step=101, value=20)
     gen.AddScalarTensor('s1', wall_time=1, step=200, value=20)
@@ -268,9 +270,9 @@ class MockingEventAccumulatorTest(EventAccumulatorTest):
     gen = _EventGenerator(self)
     acc = ea.EventAccumulator(gen)
     gen.AddScalarTensor('s1', wall_time=1, step=100, value=20)
-    ev1 = tf.Event(wall_time=2, step=0, file_version='brain.Event:1')
+    ev1 = Event(wall_time=2, step=0, file_version='brain.Event:1')
     graph_bytes = tf.GraphDef().SerializeToString()
-    ev2 = tf.Event(wall_time=3, step=0, graph_def=graph_bytes)
+    ev2 = Event(wall_time=3, step=0, graph_def=graph_bytes)
     gen.AddEvent(ev1)
     gen.AddEvent(ev2)
     acc.Reload()
@@ -285,7 +287,7 @@ class MockingEventAccumulatorTest(EventAccumulatorTest):
     """
     gen = _EventGenerator(self)
     acc = ea.EventAccumulator(gen)
-    gen.AddEvent(tf.Event(wall_time=0, step=1, file_version='brain.Event:2'))
+    gen.AddEvent(Event(wall_time=0, step=1, file_version='brain.Event:2'))
 
     gen.AddScalarTensor('s1', wall_time=1, step=100, value=20)
     gen.AddScalarTensor('s1', wall_time=1, step=200, value=20)
@@ -295,8 +297,8 @@ class MockingEventAccumulatorTest(EventAccumulatorTest):
     gen.AddScalarTensor('s2', wall_time=1, step=202, value=20)
     gen.AddScalarTensor('s2', wall_time=1, step=203, value=20)
 
-    slog = tf.SessionLog(status=tf.SessionLog.START)
-    gen.AddEvent(tf.Event(wall_time=2, step=201, session_log=slog))
+    slog = SessionLog(status=SessionLog.START)
+    gen.AddEvent(Event(wall_time=2, step=201, session_log=slog))
     acc.Reload()
     self.assertEqual([x.step for x in acc.Tensors('s1')], [100, 200])
     self.assertEqual([x.step for x in acc.Tensors('s2')], [])
@@ -305,7 +307,7 @@ class MockingEventAccumulatorTest(EventAccumulatorTest):
     """Test that FirstEventTimestamp() returns wall_time of the first event."""
     gen = _EventGenerator(self)
     acc = ea.EventAccumulator(gen)
-    gen.AddEvent(tf.Event(wall_time=10, step=20, file_version='brain.Event:2'))
+    gen.AddEvent(Event(wall_time=10, step=20, file_version='brain.Event:2'))
     gen.AddScalarTensor('s1', wall_time=30, step=40, value=20)
     self.assertEqual(acc.FirstEventTimestamp(), 10)
 
@@ -313,7 +315,7 @@ class MockingEventAccumulatorTest(EventAccumulatorTest):
     """Test that Reload() means FirstEventTimestamp() won't load events."""
     gen = _EventGenerator(self)
     acc = ea.EventAccumulator(gen)
-    gen.AddEvent(tf.Event(wall_time=1, step=2, file_version='brain.Event:2'))
+    gen.AddEvent(Event(wall_time=1, step=2, file_version='brain.Event:2'))
 
     acc.Reload()
 
@@ -327,7 +329,7 @@ class MockingEventAccumulatorTest(EventAccumulatorTest):
     """Test that FirstEventTimestamp() doesn't discard the loaded event."""
     gen = _EventGenerator(self)
     acc = ea.EventAccumulator(gen)
-    gen.AddEvent(tf.Event(wall_time=1, step=2, file_version='brain.Event:2'))
+    gen.AddEvent(Event(wall_time=1, step=2, file_version='brain.Event:2'))
 
     self.assertEqual(acc.FirstEventTimestamp(), 1)
     acc.Reload()
@@ -474,8 +476,8 @@ class MockingEventAccumulatorTest(EventAccumulatorTest):
     writer = test_util.FileWriter(self.get_temp_dir())
     writer.event_writer = event_sink
     with self.test_session() as sess:
-      summary_metadata = tf.SummaryMetadata(
-          plugin_data=tf.SummaryMetadata.PluginData(plugin_name=plugin_name,
+      summary_metadata = SummaryMetadata(
+          plugin_data=SummaryMetadata.PluginData(plugin_name=plugin_name,
                                                     content=b'{}'))
       tf.summary.tensor_summary('scalar', tf.constant(1.0),
                                 summary_metadata=summary_metadata)
@@ -534,8 +536,8 @@ class RealisticEventAccumulatorTest(EventAccumulatorTest):
     """Test accumulator by writing values and then reading them."""
 
     def FakeScalarSummary(tag, value):
-      value = tf.Summary.Value(tag=tag, simple_value=value)
-      summary = tf.Summary(value=[value])
+      value = Summary.Value(tag=tag, simple_value=value)
+      summary = Summary(value=[value])
       return summary
 
     directory = os.path.join(self.get_temp_dir(), 'values_dir')
@@ -647,7 +649,7 @@ class RealisticEventAccumulatorTest(EventAccumulatorTest):
         same file
     """
 
-    summary = tf.Summary()
+    summary = Summary()
     summary.value.add(
         tensor=tensor_util.make_tensor_proto(['po', 'ta', 'to'], dtype=tf.string),
         tag='you_are_it',
@@ -658,10 +660,10 @@ class RealisticEventAccumulatorTest(EventAccumulatorTest):
 
   def testSummaryMetadata(self):
     logdir = self.get_temp_dir()
-    summary_metadata = tf.SummaryMetadata(
+    summary_metadata = SummaryMetadata(
         display_name='current tagee',
         summary_description='no',
-        plugin_data=tf.SummaryMetadata.PluginData(plugin_name='outlet'))
+        plugin_data=SummaryMetadata.PluginData(plugin_name='outlet'))
     self._writeMetadata(logdir, summary_metadata)
     acc = ea.EventAccumulator(logdir)
     acc.Reload()
@@ -670,18 +672,18 @@ class RealisticEventAccumulatorTest(EventAccumulatorTest):
 
   def testSummaryMetadata_FirstMetadataWins(self):
     logdir = self.get_temp_dir()
-    summary_metadata_1 = tf.SummaryMetadata(
+    summary_metadata_1 = SummaryMetadata(
         display_name='current tagee',
         summary_description='no',
-        plugin_data=tf.SummaryMetadata.PluginData(plugin_name='outlet',
+        plugin_data=SummaryMetadata.PluginData(plugin_name='outlet',
                                                   content=b'120v'))
     self._writeMetadata(logdir, summary_metadata_1, nonce='1')
     acc = ea.EventAccumulator(logdir)
     acc.Reload()
-    summary_metadata_2 = tf.SummaryMetadata(
+    summary_metadata_2 = SummaryMetadata(
         display_name='tagee of the future',
         summary_description='definitely not',
-        plugin_data=tf.SummaryMetadata.PluginData(plugin_name='plug',
+        plugin_data=SummaryMetadata.PluginData(plugin_name='plug',
                                                   content=b'110v'))
     self._writeMetadata(logdir, summary_metadata_2, nonce='2')
     acc.Reload()
@@ -694,18 +696,18 @@ class RealisticEventAccumulatorTest(EventAccumulatorTest):
     # set of plugins in the `plugin_data` of second is different from
     # that of the first, then the second set should be ignored.
     logdir = self.get_temp_dir()
-    summary_metadata_1 = tf.SummaryMetadata(
+    summary_metadata_1 = SummaryMetadata(
         display_name='current tagee',
         summary_description='no',
-        plugin_data=tf.SummaryMetadata.PluginData(plugin_name='outlet',
+        plugin_data=SummaryMetadata.PluginData(plugin_name='outlet',
                                                   content=b'120v'))
     self._writeMetadata(logdir, summary_metadata_1, nonce='1')
     acc = ea.EventAccumulator(logdir)
     acc.Reload()
-    summary_metadata_2 = tf.SummaryMetadata(
+    summary_metadata_2 = SummaryMetadata(
         display_name='tagee of the future',
         summary_description='definitely not',
-        plugin_data=tf.SummaryMetadata.PluginData(plugin_name='plug',
+        plugin_data=SummaryMetadata.PluginData(plugin_name='plug',
                                                   content=b'110v'))
     self._writeMetadata(logdir, summary_metadata_2, nonce='2')
     acc.Reload()
