+++ tensorboard/plugins/audio/BUILD
@@ -57,7 +57,6 @@ py_library(
         "//tensorboard:expect_tensorflow_installed",
         "//tensorboard:tf_contrib_ffmpeg",
         "//tensorboard/util:encoder",
-        "//tensorboard/util:tensor_util",
     ],
 )
 
@@ -71,7 +70,9 @@ py_test(
         ":summary",
         "//tensorboard:expect_numpy_installed",
         "//tensorboard:expect_tensorflow_installed",
+        "//tensorboard/compat/proto:protos_all_py_pb2",
         "//tensorboard/util:tensor_util",
+        "//tensorboard/util:test_util",
     ],
 )
 
@@ -96,6 +97,7 @@ py_library(
     deps = [
         ":protos_all_py_pb2",
         "//tensorboard/compat:tensorflow",
+        "//tensorboard/compat/proto:protos_all_py_pb2",
     ],
 )
