+++ tensorboard/plugins/audio/metadata.py
@@ -20,6 +20,7 @@ from __future__ import print_function
 
 from tensorboard.plugins.audio import plugin_data_pb2
 from tensorboard.compat import tf
+from tensorboard.compat.proto.summary_pb2 import SummaryMetadata
 
 
 PLUGIN_NAME = 'audio'
@@ -33,6 +34,23 @@ Encoding = plugin_data_pb2.AudioPluginData.Encoding
 
 
 def create_summary_metadata(display_name, description, encoding):
+  """Create a `SummaryMetadata` proto for audio plugin data.
+
+  Returns:
+    A `SummaryMetadata` protobuf object.
+  """
+  content = plugin_data_pb2.AudioPluginData(
+      version=PROTO_VERSION, encoding=encoding)
+  metadata = SummaryMetadata(
+      display_name=display_name,
+      summary_description=description,
+      plugin_data=SummaryMetadata.PluginData(
+          plugin_name=PLUGIN_NAME,
+          content=content.SerializeToString()))
+  return metadata
+
+
+def create_summary_metadata_v1(display_name, description, encoding):
   """Create a `tf.SummaryMetadata` proto for audio plugin data.
 
   Returns:
