+++ tensorboard/plugins/custom_scalar/summary.py
@@ -24,7 +24,7 @@ import tensorflow as tf
 
 from tensorboard.plugins.custom_scalar import layout_pb2
 from tensorboard.plugins.custom_scalar import metadata
-from tensorboard.util import tensor_util
+
 
 def op(scalars_layout, collections=None):
   """Creates a summary that contains a layout.
@@ -64,7 +64,7 @@ def pb(scalars_layout):
     A summary proto containing the layout.
   """
   assert isinstance(scalars_layout, layout_pb2.Layout)
-  tensor = tensor_util.make_tensor_proto(
+  tensor = tf.compat.v1.make_tensor_proto(
       scalars_layout.SerializeToString(), dtype=tf.string)
   summary = tf.Summary()
   summary.value.add(tag=metadata.CONFIG_SUMMARY_TAG,
