+++ tensorboard/plugins/debugger/debugger_server_test.py
@@ -30,6 +30,7 @@ from tensorboard.plugins.debugger import constants
 from tensorboard.plugins.debugger import debugger_server_lib
 from tensorboard.plugins.debugger import numerics_alert
 from tensorboard.util import tensor_util
+from tensorboard.compat.proto.event_pb2 import Event
 # pylint: enable=ungrouped-imports, wrong-import-order
 
 
@@ -71,7 +72,7 @@ class DebuggerDataServerTest(tf.test.TestCase):
     events_writer_manager = FakeEventsWriterManager(self.events_written)
     self.stream_handler = debugger_server_lib.DebuggerDataStreamHandler(
         events_writer_manager=events_writer_manager)
-    self.stream_handler.on_core_metadata_event(tf.Event())
+    self.stream_handler.on_core_metadata_event(Event())
 
   def tearDown(self):
     tf.test.mock.patch.stopall()
@@ -87,10 +88,10 @@ class DebuggerDataServerTest(tf.test.TestCase):
       debug_op: The name of the debug op to use.
       list_of_values: A python list of values within the tensor.
     Returns:
-      A `tf.Event` with a summary containing that node name and a float64
+      A `Event` with a summary containing that node name and a float64
       tensor with those values.
     """
-    event = tf.Event()
+    event = Event()
     value = event.summary.value.add(
         tag=node_name,
         node_name="%s:%d:%s" % (node_name, output_slot, debug_op),
@@ -161,7 +162,7 @@ class DebuggerDataServerTest(tf.test.TestCase):
 
   def testCorrectStepIsWritten(self):
     events_written = []
-    metadata_event = tf.Event()
+    metadata_event = Event()
     metadata_event.log_message.message = json.dumps({"session_run_index": 42})
     stream_handler = debugger_server_lib.DebuggerDataStreamHandler(
         events_writer_manager=FakeEventsWriterManager(events_written))
@@ -179,7 +180,7 @@ class DebuggerDataServerTest(tf.test.TestCase):
 
   def testSentinelStepValueAssignedWhenExecutorStepCountKeyIsMissing(self):
     events_written = []
-    metadata_event = tf.Event()
+    metadata_event = Event()
     metadata_event.log_message.message = json.dumps({})
     stream_handler = debugger_server_lib.DebuggerDataStreamHandler(
         events_writer_manager=FakeEventsWriterManager(events_written))
@@ -191,7 +192,7 @@ class DebuggerDataServerTest(tf.test.TestCase):
 
   def testSentinelStepValueAssignedWhenMetadataJsonIsInvalid(self):
     events_written = []
-    metadata_event = tf.Event()
+    metadata_event = Event()
     metadata_event.log_message.message = "some invalid JSON string"
     stream_handler = debugger_server_lib.DebuggerDataStreamHandler(
         events_writer_manager=FakeEventsWriterManager(events_written))
@@ -207,7 +208,7 @@ class DebuggerDataServerTest(tf.test.TestCase):
         events_writer_manager=FakeEventsWriterManager(
             self.events_written),
         numerics_alert_callback=numerics_alert_callback)
-    stream_handler.on_core_metadata_event(tf.Event())
+    stream_handler.on_core_metadata_event(Event())
 
     # The stream handler receives 1 good event and 1 with an NaN value.
     stream_handler.on_value_event(
