+++ tensorboard/plugins/image/BUILD
@@ -67,7 +67,6 @@ py_library(
         ":metadata",
         "//tensorboard:expect_tensorflow_installed",
         "//tensorboard/util:encoder",
-        "//tensorboard/util:tensor_util",
     ],
 )
 
@@ -94,6 +93,7 @@ py_library(
     deps = [
         ":protos_all_py_pb2",
         "//tensorboard/compat:tensorflow",
+        "//tensorboard/compat/proto:protos_all_py_pb2",
     ],
 )
