+++ tensorboard/plugins/image/summary.py
@@ -30,7 +30,6 @@ import tensorflow as tf
 
 from tensorboard.plugins.image import metadata
 from tensorboard.util import encoder
-from tensorboard.util import tensor_util
 
 
 def op(name,
@@ -120,11 +119,11 @@ def pb(name, images, max_outputs=3, display_name=None, description=None):
   encoded_images = [encoder.encode_png(image) for image in limited_images]
   (width, height) = (images.shape[2], images.shape[1])
   content = [str(width), str(height)] + encoded_images
-  tensor = tensor_util.make_tensor_proto(content, dtype=tf.string)
+  tensor = tf.compat.v1.make_tensor_proto(content, dtype=tf.string)
 
   if display_name is None:
     display_name = name
-  summary_metadata = metadata.create_summary_metadata(
+  summary_metadata = metadata.create_summary_metadata_v1(
       display_name=display_name, description=description)
 
   summary = tf.Summary()
