+++ tensorboard/plugins/text/summary_test.py
@@ -23,9 +23,11 @@ import numpy as np
 import six
 import tensorflow as tf
 
+from tensorboard.compat.proto.summary_pb2 import Summary
 from tensorboard.plugins.text import metadata
 from tensorboard.plugins.text import summary
 from tensorboard.util import tensor_util
+from tensorboard.util import test_util
 
 
 class SummaryTest(tf.test.TestCase):
@@ -33,7 +35,7 @@ class SummaryTest(tf.test.TestCase):
   def pb_via_op(self, summary_op, feed_dict=None):
     with tf.Session() as sess:
       actual_pbtxt = sess.run(summary_op, feed_dict=feed_dict or {})
-    actual_proto = tf.Summary()
+    actual_proto = Summary()
     actual_proto.ParseFromString(actual_pbtxt)
     return actual_proto
 
@@ -46,11 +48,14 @@ class SummaryTest(tf.test.TestCase):
     normalization ensures a canonical form, and should be used before
     comparing two `Summary`s for equality.
     """
-    result = tf.Summary()
+    result = Summary()
+    if not isinstance(pb, Summary):
+      pb = test_util.tf_summary_proto_to_tb_summary_proto(pb)
     result.MergeFrom(pb)
     for value in result.value:
       if value.HasField('tensor'):
-        new_tensor = tensor_util.make_tensor_proto(tensor_util.make_ndarray(value.tensor))
+        new_tensor = tensor_util.make_tensor_proto(
+            tensor_util.make_ndarray(value.tensor))
         value.ClearField('tensor')
         value.tensor.MergeFrom(new_tensor)
     return result
